Pycolab maze environment implemented, and integrated with the codebase. Runs with env-id = maze.
This is a simpler environment that should be easily modifiable, and should provide high FPS and much faster training times for debugging.
Pycolab provides the environment itself (better_scrolly_maze), and mazeworld is simply an OpenAI gym interface for this environment. When defining new environments, define them clearly in Pycolab and then connect/export them appropriately throughout mazeworld.
External reward is working, but there was a small bug with dying when eaten by an enemy. In order to fix this, we were sure to put the player as the first sprite to update. That way, the players position is fixed when the enemy positions are updated. It is when enemy positions are updated that player death is checked.